Move WP_Rocket\Subscriber\CDN\RocketCDN\ to WP_Rocket\Engine\CDN\RocketCDN\ and move/adjust the ServiceProvider.php
