+++ tests/Integration/inc/classes/subscriber/CDN/RocketCDN/AdminPageSubscriber/addSubscriptionModal.php
@@ -5,7 +5,7 @@ namespace WP_Rocket\Tests\Integration\inc\classes\subscriber\CDN\RocketCDN;
 use WPMedia\PHPUnit\Integration\TestCase;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\AdminPageSubscriber::add_subscription_modal
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\AdminPageSubscriber::add_subscription_modal
  * @group  RocketCDN
  * @group  AdminOnly
  */
