+++ tests/Integration/inc/classes/subscriber/CDN/RocketCDN/NoticesSubscriber/addDismissScript.php
@@ -5,7 +5,7 @@ namespace WP_Rocket\Tests\Integration\inc\classes\subscriber\CDN\RocketCDN;
 use WPMedia\PHPUnit\Integration\TestCase;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\NoticesSubscriber::add_dismiss_script
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\NoticesSubscriber::add_dismiss_script
  * @group  RocketCDN
  * @group  AdminOnly
  */
