+++ tests/Integration/inc/classes/subscriber/CDN/RocketCDN/RESTSubscriber/disable.php
@@ -5,7 +5,7 @@ namespace WP_Rocket\Tests\Integration\inc\classes\subscriber\CDN\RocketCDN\RESTS
 use WP_Rocket\Tests\Integration\ApiTestCase;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\RESTSubscriber::disable
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\RESTSubscriber::disable
  * @group  RocketCDN
  */
 class Test_Disable extends ApiTestCase {
