+++ tests/Integration/inc/classes/subscriber/CDN/RocketCDN/RESTSubscriber/registerEnableRoute.php
@@ -6,9 +6,9 @@ use WP_Rocket\Tests\Integration\ApiTestCase;
 use WPMedia\PHPUnit\Integration\TestCase;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\RESTSubscriber::register_enable_route
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\RESTSubscriber::validateEmail
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\RESTSubscriber::validateKey
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\RESTSubscriber::register_enable_route
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\RESTSubscriber::validateEmail
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\RESTSubscriber::validateKey
  * @group  RocketCDN
  */
 class Test_RegisterEnableRoute extends ApiTestCase {
