+++ tests/Unit/inc/classes/subscriber/CDN/RocketCDN/AdminPageSubscriber/displayManageSubscription.php
@@ -7,10 +7,10 @@ use WPMedia\PHPUnit\Unit\TestCase;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Admin\Beacon\Beacon;
 use WP_Rocket\Engine\CDN\RocketCDN\APIClient;
-use WP_Rocket\Subscriber\CDN\RocketCDN\AdminPageSubscriber;
+use WP_Rocket\Engine\CDN\RocketCDN\AdminPageSubscriber;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\AdminPageSubscriber::display_manage_subscription
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\AdminPageSubscriber::display_manage_subscription
  * @group  RocketCDN
  */
 class Test_DisplayManageSubscription extends TestCase {
