+++ tests/Unit/inc/classes/subscriber/CDN/RocketCDN/NoticesSubscriber/promoteRocketcdnNotice.php
@@ -5,11 +5,11 @@ namespace WP_Rocket\Tests\Unit\inc\classes\subscriber\CDN\RocketCDN;
 use Brain\Monkey\Functions;
 use Mockery;
 use WP_Rocket\Engine\CDN\RocketCDN\APIClient;
-use WP_Rocket\Subscriber\CDN\RocketCDN\NoticesSubscriber;
+use WP_Rocket\Engine\CDN\RocketCDN\NoticesSubscriber;
 use WPMedia\PHPUnit\Unit\TestCase;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\RocketCDN\NoticesSubscriber::promote_rocketcdn_notice
+ * @covers \WP_Rocket\Engine\CDN\RocketCDN\NoticesSubscriber::promote_rocketcdn_notice
  * @group  RocketCDN
  */
 class Test_PromoteRocketcdnNotice extends TestCase {
